fix(security): move the JSX artifact retention into the cache module
Review of the previous revision raised the placement of the cleanup this
change adds and three narrower defects, all valid.

Move the artifact retention out of the transform module. The served and
mtime memos, the per-artifact lock, the reference counts a render holds,
the scheduled prune passes and the prune itself now live in `jsx-cache.ts`
next to `ensureCachedJsxModulePatched`, which already owned the cached
artifact's normalization. `import-transformer.ts` keeps only the transform
and its source admission, and its test-only export drops from twenty keys
to four.

Read the project source through the repo's shared bounded reader.
`captureBoundedTextReader` captures the adapter's capabilities without
invoking accessors or Proxy traps, re-checks the returned byte length so a
reader that hands back more than its ceiling is refused rather than
admitted, and decodes strict UTF-8. The prefix-reader fallback admits its
bytes through the same fixed copy. Only a filesystem advertising neither
an exact reader nor a small enough whole-file ceiling reaches `readFile`.

Require a retained artifact to sit inside the cache directory. Selection
was by name shape alone, so a tenant-authored `file://` import of an
`jsx-*.mjs` path anywhere on disk was pinned, given a real `utime` write
on a path this cache does not own, and allowed to evict genuine marks from
the served memo.

Bound the directories holding a scheduled prune, and drop the separate
armed-sweep set that grew with every cache directory the process ever
touched. Arming is now keyed on whether the directory already has a
pending pass, so a process serving only cache hits keeps rescanning
instead of arming once and never looking again.
